split standardize_mutation_data.py to vcf2maf and maf2maf tools
